Move invalidation methods into a module
